MSAPI-80 | Rework Document Tests
