+++ SignNow.Net.Test/AcceptanceTests/DocumentServiceTest.Delete.cs
@@ -1,8 +1,6 @@
 using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SignNow.Net.Exceptions;
 using SignNow.Net.Test;
-using System.IO;
 using System.Threading.Tasks;
 using System.Text.RegularExpressions;
 
@@ -13,9 +11,9 @@ namespace AcceptanceTests
         [TestMethod]
         public void DocumentDeletingSuccess()
         {
-            var testDocumentId = UploadTestDocument(PdfFilePath);
+            DocumentId = UploadTestDocument(PdfFilePath);
 
-            var deleteResponse = docService.DeleteDocumentAsync(testDocumentId);
+            var deleteResponse = docService.DeleteDocumentAsync(DocumentId);
             Task.WaitAll(deleteResponse);
 
             Assert.IsFalse(
@@ -49,29 +47,9 @@ namespace AcceptanceTests
             {
                 foreach (var exception in ex.InnerExceptions)
                 {
-                    Console.WriteLine(documentId);
                     Assert.AreEqual(message, exception.Message);
-
                 }
             }
-        }
-
-        private string UploadTestDocument(string filePath)
-        {
-            string docId = default;
-
-            using (var fileStream = File.OpenRead(filePath))
-            {
-                var uploadResponse = docService.UploadDocumentAsync(fileStream, pdfFileName).Result;
-                docId = uploadResponse.Id;
-
-                Assert.IsNotNull(
-                    uploadResponse.Id,
-                    "Document Upload result should contain non-null Id property value on successful upload"
-                    );
-            }
-
-            return docId;
         }
     }
 }
