+++ SignNow.Net.Test/AcceptanceTests/DocumentServiceTest.Upload.cs
@@ -7,7 +7,6 @@ using SignNow.Net.Test.SignNow;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -29,7 +28,6 @@ namespace AcceptanceTests
         {
             DocumentUploadTestServices(DocumentUploadTest, d => d.UploadDocumentWithFieldExtractAsync);
             //TODO: test if fields were extracted correctly
-
         }
 
         [TestMethod]
@@ -59,6 +57,7 @@ namespace AcceptanceTests
                 {
                     var uploadResponse = uploadFunction(fileStream, pdfFileName, default).Result;
                     docId = uploadResponse.Id;
+
                     Assert.IsNotNull(uploadResponse.Id, "Document Upload result should contain non-null Id property value on successful upload");
                 }
                 finally
@@ -77,33 +76,16 @@ namespace AcceptanceTests
                 {
                     var uploadResponse = uploadFunction(fileStream, txtFileName,default).Result;
                     docId = uploadResponse.Id;
-
                 }
                 catch (AggregateException ex)
                 {
-                    
                     Assert.AreEqual(ErrorMessages.InvalidFileType, ex.InnerException.Message);
                 }
                 finally
                 {
                     DeleteDocument(docId);
                 }
-
             }
-        }
-        void DeleteDocument(string id)
-        {
-            if (string.IsNullOrEmpty(id))
-                return;
-            using (var client = new HttpClient())
-            using (var requestMessage = new HttpRequestMessage(HttpMethod.Delete, $"{ApiBaseUrl}/document/{id}"))
-            {
-                requestMessage.Headers.Add("Authorization", Token.GetAuthorizationHeaderValue());
-                var response = client.SendAsync(requestMessage).Result;
-                response.EnsureSuccessStatusCode();
-            }
-                
-            //docService.DeleteDocumentAsync(id).RunSynchronously();
         }
     }
 }
